+++ NEWS.md
@@ -1,5 +1,6 @@
 # occumb (development version)
 * Add `predict()` method for `occumbFit` class.
+* Fix `predict()` to handle cases where some covariate type is missing.
 
 ## occumb 1.0.3 (2024/01/04)
 * Internal changes to fix a number of known bugs and for more helpful error messages and additional tests.